# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  title: Api
  version: 0.1.0
  description: TiliGo Delivery Platform API
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /products:
    get:
      operationId: listProducts
      tags: [products]
      parameters:
        - in: query
          name: business_id
          schema:
            type: integer
        - in: query
          name: is_available
          schema:
            type: boolean
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Product"
    post:
      operationId: createProduct
      tags: [products]
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ProductInput"
      responses:
        "201":
          description: Created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Product"
  /products/{id}:
    patch:
      operationId: updateProduct
      tags: [products]
      parameters:
        - in: path
          name: id
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ProductUpdate"
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Product"
    delete:
      operationId: deleteProduct
      tags: [products]
      parameters:
        - in: path
          name: id
          required: true
          schema:
            type: integer
      responses:
        "204":
          description: Deleted

  # ─── OFFERS ────────────────────────────────────────────────────────
...
  /tickets:
    get:
      operationId: listTickets
      tags: [tickets]
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/SupportTicket"
    post:
      operationId: createTicket
      tags: [tickets]
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/SupportTicketInput"
      responses:
        "201":
          description: Created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/SupportTicket"
  /tickets/{id}:
    patch:
      operationId: updateTicket
      tags: [tickets]
      parameters:
        - in: path
          name: id
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/SupportTicketUpdate"
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/SupportTicket"
    delete:
      operationId: deleteTicket
      tags: [tickets]
      parameters:
        - in: path
          name: id
          required: true
          schema:
            type: integer
      responses:
        "204":
          description: Deleted

  # ─── DASHBOARD ─────────────────────────────────────────────────────
...
components:
  schemas:
...
    Product:
      type: object
      properties:
        id:           { type: integer }
        business_id:  { type: integer }
        business_name: { type: ["string", "null"] }
        name:         { type: string }
        description:  { type: ["string", "null"] }
        price:        { type: number }
        category:     { type: ["string", "null"] }
        image_url:    { type: ["string", "null"] }
        is_available: { type: boolean }
        created_at:   { type: string }
      required: [id, business_id, name, price, is_available]
    ProductInput:
      type: object
      properties:
        business_id:  { type: integer }
        business_name: { type: string }
        name:         { type: string }
        description:  { type: string }
        price:        { type: number }
        category:     { type: string }
        image_url:    { type: string }
        is_available: { type: boolean }
      required: [business_id, name, price]
    ProductUpdate:
      type: object
      properties:
        name:         { type: string }
        description:  { type: string }
        price:        { type: number }
        category:     { type: string }
        image_url:    { type: string }
        is_available: { type: boolean }

    # ── Offer ─────────────────────────────────────────────────────────
...
    SupportTicket:
      type: object
      properties:
        id:         { type: integer }
        name:       { type: string }
        email:      { type: string }
        phone:      { type: ["string", "null"] }
        subject:    { type: string }
        message:    { type: string }
        category:   { type: ["string", "null"] }
        priority:   { type: string }
        status:     { type: string }
        created_at: { type: string }
      required: [id, name, email, subject, message, priority, status]
    SupportTicketInput:
      type: object
      properties:
        name:     { type: string }
        email:    { type: string }
        phone:    { type: string }
        subject:  { type: string }
        message:  { type: string }
        category: { type: string }
        priority: { type: string }
      required: [name, email, subject, message]
    SupportTicketUpdate:
      type: object
      properties:
        status:   { type: string }
        priority: { type: string }

    # ── Dashboard ─────────────────────────────────────────────────────
...