# lib/api-zod/src/generated/types/product.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * TiliGo Delivery Platform API
 * OpenAPI spec version: 0.1.0
 */

export interface Product {
  id: number;
  business_id: number;
  /** @nullable */
  business_name?: string | null;
  name: string;
  /** @nullable */
  description?: string | null;
  price: number;
  /** @nullable */
  category?: string | null;
  /** @nullable */
  image_url?: string | null;
  is_available: boolean;
  created_at?: string;
}