# lib/api-zod/src/generated/types/productInput.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * TiliGo Delivery Platform API
 * OpenAPI spec version: 0.1.0
 */

export interface ProductInput {
  business_id: number;
  business_name?: string;
  name: string;
  description?: string;
  price: number;
  category?: string;
  image_url?: string;
  is_available?: boolean;
}