# lib/api-zod/src/generated/types/supportTicket.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * TiliGo Delivery Platform API
 * OpenAPI spec version: 0.1.0
 */

export interface SupportTicket {
  id: number;
  name: string;
  email: string;
  /** @nullable */
  phone?: string | null;
  subject: string;
  message: string;
  /** @nullable */
  category?: string | null;
  priority: string;
  status: string;
  created_at?: string;
}