# lib/api-zod/src/generated/types/supportTicketUpdate.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * TiliGo Delivery Platform API
 * OpenAPI spec version: 0.1.0
 */

export interface SupportTicketUpdate {
  status?: string;
  priority?: string;
}